클래스끼리 묶어 따로 관리
