+++ classes/Help.py
@@ -1,19 +1,19 @@
-from constants.COMMAND import BASE, HELP, IDENTIFIER      # 상수
-from command.commands import Commands
+from classes import Commands
+from constants.COMMAND import HELP, BASE
 
 
 class Help(Commands):     # 도움 커맨드
     def __init__(self, message, say):
         Commands.__init__(self, message, say)
 
-    def errorCheck(self):
+    def checkError(self):
         messageToken = self.message.text.split()
         # 설마 명령어가 잘못되지는 않았겠지?
-        if self.message.command != HELP.COMMAND_IDENTIFIER[HELP.COMMAND]:
+        if self.message.command != self.identifier:
             return False
         return True if len(messageToken) == 1 else False
 
-    def getTextForChannel(self):
+    def getHelpMessage(self):
         output = HELP.HELP_HEADLINE
 
         # 출력할 리스트에서 하나씩 "명령어 : 도움말" 형식으로 출력 진행
@@ -23,17 +23,12 @@ class Help(Commands):     # 도움 커맨드
 
         return output
 
+    '''
+        식별자와 커맨드명을 리턴해 주는 함수
+    '''
 
-'''
-    커맨드 체크 명령어 : manageCommand에서 커맨드 분류에 사용
-'''
+    def getCommandName(command):
+        return HELP.COMMAND_NAME[command]
 
-
-def isHelpCommand(message):  # help 명령어인지 확인하는 함수
-    # help 명령어 인지 확인
-    return True if message.text.startswith(IDENTIFIER.HELP) else False
-
-
-def controlHelpCommand(message, say):      # 도움 관련 명령어 제어문
-    works = Help(message, say)
-    works.sayCommand()      # 커맨드 출력
+    def getIdentifier(command):
+        return HELP.COMMAND_IDENTIFIER[command]
